CD-ROM images are now working correctly again; Fixed all the reported bugs regarding the Settings dialog; MIDI out device is now no longer reset to 0 after hard reset; Removed all vestiges of the old disk activity flash; The configuration file is no longer saved when it shouldn't be; Redone the status bar icon updating so it is only done in win.c; Made sure all variables in ibm.h are extern; A lot of other bugfixes; Added Mouse Systems Mouse emulation (patch from TheCollector1995); Added IBM PS/1 Model 2133 (486) emulation (patch from TheCollector1995); Tweaked the CPU dynamic recompiler cycle periods - 486SX 33 and 486DX 33 now work; Increased compatibility with configuration files from before the previous commit.

+++ src/cdrom_dosbox.h
@@ -103,7 +103,7 @@ class CDROM_Interface_Image : public CDROM_Interface
 private:
 	class TrackFile {
 	public:
-		virtual bool read(Bit8u *buffer, int seek, int count) = 0;
+		virtual bool read(Bit8u *buffer, uint64_t seek, uint64_t count) = 0;
 		virtual uint64_t getLength() = 0;
 		virtual ~TrackFile() { };
 	};
@@ -112,7 +112,7 @@ private:
 	public:
 		BinaryFile(const char *filename, bool &error);
 		~BinaryFile();
-		bool read(Bit8u *buffer, int seek, int count);
+		bool read(Bit8u *buffer, uint64_t seek, uint64_t count);
 		uint64_t getLength();
 	private:
 		BinaryFile();
